requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
PyJWT==2.8.0
pandas==2.0.3
openpyxl==3.1.2